[ADF-248] [ADF-249] Update Logic

+++ BusinessObjects/DTOs/UserDTOs/UserTokenDTO.cs
@@ -34,7 +34,7 @@ namespace BusinessObjects
 
         public int Role { get; set; }
 
-        public int Wallet { get; set; }
+        public decimal Wallet { get; set; }
 
         public int Provider { get; set; }
